Move status API methods into trait that can be mixed into SparkContext.
